Small changes in dbprop handling.

+++ trunk/components-ext/src/main/java/org/dllearner/algorithm/tbsl/sparql/Allocation.java
@@ -10,7 +10,7 @@ public class Allocation {
 	
 	private double score;
 	
-	public Allocation(String uri, int prominence, double similarity) {
+	public Allocation(String uri, double prominence, double similarity) {
 		this.uri = uri;
 		this.prominence = prominence;
 		this.similarity = similarity;
